try to fix multi app test
