+++ springboot-samples/logging/log4j2/log4j2-integration-test/pom.xml
