+++ springboot-samples/pom.xml
@@ -14,7 +14,7 @@
         <java.version>1.8</java.version>
         <maven.compiler.source>8</maven.compiler.source>
         <maven.compiler.target>8</maven.compiler.target>
-        <koupleless.runtime.version>1.2.2</koupleless.runtime.version>
+        <koupleless.runtime.version>1.2.3-SNAPSHOT</koupleless.runtime.version>
         <sofa.ark.version>2.2.11</sofa.ark.version>
         <disruptor.version>3.4.2</disruptor.version>
         <os.plugin.version>1.7.1</os.plugin.version>
