+++ springboot-samples/web/tomcat/base-web-single-host/base-web-single-host-bootstrap/Formatter.xml
