+++ springboot-samples/web/tomcat/base-web-single-host/base-web-single-host-bootstrap/src/main/java/com/alipay/sofa/web/base/BaseApplication.java
