+++ springboot-samples/web/tomcat/base-web-single-host/facade/pom.xml
