+++ springboot-samples/web/tomcat/base-web-single-host/facade/src/main/java/com/alipay/sofa/dynamicstock/base/model/ProductInfo.java
