+++ springboot-samples/web/tomcat/biz1-web-single-host/src/main/java/com/alipay/sofa/web/biz1/Biz1Application.java
