+++ springboot-samples/web/tomcat/biz1-web-single-host/src/main/resources/config/application.properties
