+++ springboot-samples/web/tomcat/biz2-web-single-host/src/main/java/com/alipay/sofa/web/biz2/Biz2Application.java
