+++ springboot-samples/web/tomcat/biz2-web-single-host/src/main/resources/application.properties
