--- springboot-samples/web/tomcat/biz2/src/main/java/com/alipay/sofa/web/biz2/rest/SampleController.java
@@ -1,46 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.alipay.sofa.web.biz2.rest;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.ApplicationContext;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.RestController;
-
-import java.time.LocalDate;
-import java.time.LocalDateTime;
-
-@RestController
-public class SampleController {
-
-    @Autowired
-    private ApplicationContext applicationContext;
-
-    @RequestMapping(value = "/", method = RequestMethod.GET)
-    public String hello() {
-        String appName = applicationContext.getId();
-        return String.format("hello to %s deploy, run in %s", appName, Thread.currentThread()
-            .getContextClassLoader().getClass().getName());
-    }
-
-    @RequestMapping(value = "/timestamp", method = RequestMethod.GET)
-    public String timestamp() {
-        String appName = applicationContext.getApplicationName();
-        return String.format("%s now is %s", appName, LocalDateTime.now());
-    }
-}
