+++ springboot-samples/web/tomcat/tomcat-web-integration-test/src/test/java/com/alipay/sofa/web/webflux/integration/test/WebSingleHostTest.java
@@ -14,10 +14,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.alipay.sofa.web.integration.test;
+package com.alipay.sofa.web.webflux.integration.test;
 
-import com.alipay.sofa.dynamicstock.base.facade.StrategyService;
-import com.alipay.sofa.koupleless.common.api.SpringServiceFinder;
 import com.alipay.sofa.koupleless.test.suite.spring.model.BaseSpringTestConfig;
 import com.alipay.sofa.koupleless.test.suite.spring.model.BizSpringTestConfig;
 import com.alipay.sofa.koupleless.test.suite.spring.model.MultiSpringTestConfig;
@@ -27,6 +25,7 @@ import com.alipay.sofa.web.biz2.Biz2Application;
 import com.alipay.sofa.web.base.BaseApplication;
 import com.google.common.collect.Lists;
 import lombok.SneakyThrows;
+import lombok.extern.slf4j.Slf4j;
 import okhttp3.OkHttpClient;
 import okhttp3.Request;
 import okhttp3.Response;
@@ -38,6 +37,7 @@ import org.junit.Test;
  * @author CodeNoobKing
  * @date 2024/3/11
  */
+@Slf4j
 public class WebSingleHostTest {
 
     private static TestMultiSpringApplication multiApp;
@@ -64,17 +64,20 @@ public class WebSingleHostTest {
     public void testContextWebhookPathPrefixIsAdded() throws Throwable {
         Response resp = client.newCall(
             new Request.Builder().url("http://localhost:8080/").get().build()).execute();
-        Assert.assertEquals("hello to base deploy", resp.body().string());
+        String result = resp.body().string();
+        log.info("result: {}", result);
+        Assert.assertTrue(result.contains("hello to base deploy"));
 
         resp = client.newCall(
             new Request.Builder().url("http://localhost:8080/biz1/").get().build()).execute();
-        Assert.assertEquals("hello to biz1 deploy", resp.body().string());
+        result = resp.body().string();
+        log.info("result: {}", result);
+        Assert.assertTrue(result.contains("hello to biz1 deploy"));
 
         resp = client.newCall(
             new Request.Builder().url("http://localhost:8080/biz2/").get().build()).execute();
-        Assert.assertEquals("hello to biz2 deploy", resp.body().string());
-
-        Assert.assertEquals("biz1",
-            SpringServiceFinder.getModuleService("biz1", null, StrategyService.class).getAppName());
+        result = resp.body().string();
+        log.info("result: {}", result);
+        Assert.assertTrue(result.contains("hello to biz2 deploy"));
     }
 }
